+++ src/main/java/by/it/academy/controllers/LogoutController.java
@@ -1,0 +1,20 @@
+package by.it.academy.controllers;
+
+import javax.servlet.annotation.WebServlet;
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
+import java.io.IOException;
+
+@WebServlet(urlPatterns = {"/logout"})
+public class LogoutController extends HttpServlet {
+    private static final long serialVersionUID = 344602554021L;
+
+    @Override
+    protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws IOException {
+        HttpSession session = req.getSession();
+        session.removeAttribute("userType");
+        resp.sendRedirect("/index.jsp");
+    }
+}
