+++ src/main/webapp/index.jsp
@@ -19,8 +19,9 @@
 </div>
 <div class="w3-container w3-center">
     <div class="w3-bar w3-padding-large w3-padding-24">
-        <button class="w3-btn w3-hover-yellow w3-round-large" onclick="location.href='/login'">Sign in</button>
-        <button class="w3-btn w3-hover-light-blue w3-round-large" onclick="location.href='/user/read'">List users</button>
+        <button class="w3-btn w3-hover-yellow w3-round-large" onclick="location.href='/login'">Login</button>
+        <button class="w3-btn w3-hover-orange w3-round-large" onclick="location.href='/logout'">Logout</button>
+        <button class="w3-btn w3-hover-light-blue w3-round-large" onclick="location.href='/user/read'">Users List</button>
         <button class="w3-btn w3-hover-green w3-round-large" onclick="location.href='/user/create'">Register</button>
         <button class="w3-btn w3-hover-red w3-round-large" onclick="location.href='/cars'">Cars list</button>
     </div>
